[Core] Move Task from TargetType.swift to its own file
