+++ Sources/Moya/TargetType.swift
@@ -30,37 +30,3 @@ public extension TargetType {
         return false
     }
 }
-
-/// Represents an HTTP task.
-public enum Task {
-
-    /// A request with no additional data.
-    case requestPlain
-
-    /// A requests body set with data.
-    case requestData(Data)
-
-    /// A requests body set with encoded parameters.
-    case requestParameters(parameters: [String: Any], encoding: ParameterEncoding)
-
-    /// A requests body set with data, combined with url parameters.
-    case requestCompositeData(bodyData: Data, urlParameters: [String: Any])
-
-    /// A requests body set with encoded parameters combined with url parameters.
-    case requestCompositeParameters(bodyParameters: [String: Any], bodyEncoding: ParameterEncoding, urlParameters: [String: Any])
-
-    /// A file upload task.
-    case uploadFile(URL)
-
-    /// A "multipart/form-data" upload task.
-    case uploadMultipart([MultipartFormData])
-
-    /// A "multipart/form-data" upload task  combined with url parameters.
-    case uploadCompositeMultipart([MultipartFormData], urlParameters: [String: Any])
-
-    /// A file download task to a destination.
-    case downloadDestination(DownloadDestination)
-
-    /// A file download task to a destination with extra parameters using the given encoding.
-    case downloadParameters(parameters: [String: Any], encoding: ParameterEncoding, destination: DownloadDestination)
-}
